# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Remote Patient Monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List all providers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
    post:
      operationId: createProvider
      tags: [providers]
      summary: Create a new provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderRequest"
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
...
  /patients/{patientId}/vitals:
    get:
      operationId: getPatientVitals
      tags: [vitals]
      summary: Get vitals history for a patient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Vitals history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VitalsReading"
    post:
      operationId: ingestVitals
      tags: [vitals]
      summary: Ingest vitals data for a patient (from wearable or manual input)
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestVitalsRequest"
      responses:
        "201":
          description: Vitals ingested and alerts evaluated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestVitalsResponse"
  /patients/{patientId}/vitals/latest:
    get:
      operationId: getLatestVitals
      tags: [vitals]
      summary: Get the most recent vitals for a patient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Latest vitals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalsReading"
        "404":
          description: No vitals found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/{patientId}/summary:
    get:
      operationId: getPatientSummary
      tags: [summary]
      summary: Get daily summary stats for a patient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient daily summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDailySummary"
  /patients/{patientId}/thresholds:
    get:
      operationId: getPatientThresholds
      tags: [thresholds]
      summary: Get alert thresholds for a patient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Threshold configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalThresholds"
    put:
      operationId: updatePatientThresholds
      tags: [thresholds]
      summary: Update alert thresholds for a patient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateThresholdsRequest"
      responses:
        "200":
          description: Thresholds updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalThresholds"
...
  /vitals/ingest-batch:
    post:
      operationId: ingestVitalsBatch
      tags: [vitals]
      summary: Batch ingest vitals for multiple patients (wearable sync)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchIngestRequest"
      responses:
        "200":
          description: Batch processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchIngestResponse"
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [summary]
      summary: Get aggregate stats for the provider dashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    Provider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        specialty:
          type: string
        patientCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - patientCount
        - createdAt
    CreateProviderRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        specialty:
          type: string
      required:
        - name
        - email
        - password
...
    VitalsReading:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        heartRate:
          type: number
          description: BPM
        systolicBp:
          type: number
          description: mmHg
        diastolicBp:
          type: number
          description: mmHg
        spo2:
          type: number
          description: Oxygen saturation percentage
        caloriesBurned:
          type: number
        temperature:
          type: number
          description: Celsius
        source:
          type: string
          enum: [apple_health, google_fit, fitbit, garmin, manual, simulated]
        recordedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - recordedAt
        - createdAt
        - source
...
    IngestVitalsRequest:
      type: object
      properties:
        heartRate:
          type: number
        systolicBp:
          type: number
        diastolicBp:
          type: number
        spo2:
          type: number
        caloriesBurned:
          type: number
        temperature:
          type: number
        source:
          type: string
          enum: [apple_health, google_fit, fitbit, garmin, manual, simulated]
        recordedAt:
          type: string
          format: date-time
      required:
        - source
    IngestVitalsResponse:
      type: object
      properties:
        vitals:
          $ref: "#/components/schemas/VitalsReading"
        alertsTriggered:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        riskLevel:
          type: string
          enum: [normal, warning, critical]
      required:
        - vitals
        - alertsTriggered
        - riskLevel
    BatchIngestRequest:
      type: object
      properties:
        readings:
          type: array
          items:
            type: object
            properties:
              patientId:
                type: integer
              vitals:
                $ref: "#/components/schemas/IngestVitalsRequest"
            required:
              - patientId
              - vitals
      required:
        - readings
    BatchIngestResponse:
      type: object
      properties:
        processed:
          type: integer
        alertsTriggered:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - processed
        - alertsTriggered
        - errors
    Alert:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        vitalType:
          type: string
          enum: [heart_rate, blood_pressure, spo2, temperature]
        severity:
          type: string
          enum: [warning, critical]
        message:
          type: string
        value:
          type: number
        threshold:
          type: number
        status:
          type: string
          enum: [active, acknowledged, resolved]
        suggestedAction:
          type: string
        triggeredAt:
          type: string
          format: date-time
        acknowledgedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - vitalType
        - severity
        - message
        - status
        - triggeredAt
...
    VitalThresholds:
      type: object
      properties:
        patientId:
          type: integer
        heartRateMin:
          type: number
          default: 50
        heartRateMax:
          type: number
          default: 100
        heartRateCriticalMin:
          type: number
          default: 40
        heartRateCriticalMax:
          type: number
          default: 120
        systolicBpMin:
          type: number
          default: 90
        systolicBpMax:
          type: number
          default: 140
        systolicBpCriticalMin:
          type: number
          default: 80
        systolicBpCriticalMax:
          type: number
          default: 180
        diastolicBpMin:
          type: number
          default: 60
        diastolicBpMax:
          type: number
          default: 90
        diastolicBpCriticalMax:
          type: number
          default: 120
        spo2Min:
          type: number
          default: 95
        spo2CriticalMin:
          type: number
          default: 90
        temperatureMin:
          type: number
          default: 36.0
        temperatureMax:
          type: number
          default: 37.5
        temperatureCriticalMax:
          type: number
          default: 39.0
      required:
        - patientId
    UpdateThresholdsRequest:
      type: object
      properties:
        heartRateMin:
          type: number
        heartRateMax:
          type: number
        heartRateCriticalMin:
          type: number
        heartRateCriticalMax:
          type: number
        systolicBpMin:
          type: number
        systolicBpMax:
          type: number
        systolicBpCriticalMin:
          type: number
        systolicBpCriticalMax:
          type: number
        diastolicBpMin:
          type: number
        diastolicBpMax:
          type: number
        diastolicBpCriticalMax:
          type: number
        spo2Min:
          type: number
        spo2CriticalMin:
          type: number
        temperatureMin:
          type: number
        temperatureMax:
          type: number
        temperatureCriticalMax:
          type: number
    PatientDailySummary:
      type: object
      properties:
        patientId:
          type: integer
        date:
          type: string
          format: date
        avgHeartRate:
          type: number
        minHeartRate:
          type: number
        maxHeartRate:
          type: number
        avgSystolicBp:
          type: number
        avgDiastolicBp:
          type: number
        avgSpo2:
          type: number
        totalCalories:
          type: number
        readingsCount:
          type: integer
        alertsCount:
          type: integer
        riskLevel:
          type: string
          enum: [normal, warning, critical]
      required:
        - patientId
        - date
        - readingsCount
        - alertsCount
        - riskLevel
    DashboardStats:
      type: object
      properties:
        totalPatients:
          type: integer
        activeAlerts:
          type: integer
        criticalPatients:
          type: integer
        warningPatients:
          type: integer
        normalPatients:
          type: integer
        todaysReadings:
          type: integer
      required:
        - totalPatients
        - activeAlerts
        - criticalPatients
        - warningPatients
        - normalPatients
        - todaysReadings