# lib/api-zod/src/generated/types/createProviderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Remote Patient Monitoring API
 * OpenAPI spec version: 1.0.0
 */

export interface CreateProviderRequest {
  name: string;
  email: string;
  password: string;
  specialty?: string;
}